Added Archetypes

+++ App/Source/HelloWorldLayer.h
@@ -8,7 +8,7 @@
 class HelloWorldLayer : public Mupfel::Layer
 {
 	void OnInit() override;
-	void OnUpdate(float timestep) override;
+	void OnUpdate(double timestep) override;
 	void OnRender() override;
 private:
 	void ProcessEvents();
